Added Try/Except Statement for USB Disconnect
When the USB cable is disconnected this try/except statement now exits
python and displays a helpful message suggesting that the USB cable has
been disconnected. This exits all of Proxy as a consequesce so it's pretty
drastic but seems reasonable. It's not guaranteed to be the USB disconnect
everytime but most likely will be. Addresses a portion of Issue #62