Updated logging for APRS and importing configs
APRS is now in loggingConfig.ini with a special APRS logger. However this
doesn't seem to actually print to the screen in debug for some reason even
though I can confirm the logger is in debug mode. Might just be PyCharm
debug messing with it. Also, I moved in code to properly look for and read
aprs.ini.