FINALLY WORKING RECEIVE OVER DATA SOCKETS
I separated the threads so we have a different socket +10 higher than RX
socket that is transmitting the receive proxy data from the RF link.
Not sure if this was 100% necessary but I did it. Then I realized I HAD to
send a byte of data (" ") to the receive socket before I could get a
response from it with the RX data from getDicts.. Wierd.