Fixed uncaught exceptions when unconfigured
Catch exceptions for REPLACEME text in callsign and nodeid. Basically
check for REPLACEME and raise an exception while returning blank data ''
from Flask and printing useful info to the logger. I also check for a
ValueError on the nodeid since one will be raised with REPLACEME in the
nodeid field.