Fixed merge issues w/ leftover configuration code
Yeah it was a messy merge. I will risk it and move forward. I also intend
to implement unit testing in the near future so we can start to be sure
stuff doesn't change unexpectedly between commits. This commit cleans up
the use of the helper class in SimpleUI with the recent merge with master
code.