Fixed Telemetry sequence counting bug
I never stopped the counter from incrementing and once the value reached
8280 and incremented to 8281 the telemetry suddenly was rejected/ignored
by aprs.fi. This led to realizing the counter is limited to this value and
I should just reset. Therefore this fix just implements a for loop
counting in the infinite while loop which resets to zero after counting to
8179. Yes I stayed one count away from the maximum just incase there is
misunderstanding about zero indexing.